Internal traces refactored
