+++ lib/ddtrace/utils.rb
@@ -1,4 +1,5 @@
 require 'ddtrace/utils/database'
+require 'ddtrace/utils/internal_traces'
 
 module Datadog
   # Utils contains low-level utilities, typically to provide pseudo-random trace IDs.
